requirements.txt
backports.ssl-match-hostname==3.4.0.2
boto==2.38.0
bz2file==0.98
certifi==2015.9.6.2
decorator==4.0.4
functools32==3.2.3.post2
gensim==0.12.2
httpretty==0.8.6
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.0.0
jupyter-core==4.0.6
lda==1.0.2
MarkupSafe==0.23
mistune==0.7.1
nbconvert==4.0.0
nbformat==4.0.0
nltk==3.0.5
notebook==4.0.5
numpy==1.9.3
path.py==8.1.1
pbr==0.11.0
pexpect==3.3
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
pyzmq==14.7.0
requests==2.7.0
scikit-learn==0.16.1
scipy==0.16.0
simplegeneric==0.8.1
six==1.9.0
sklearn==0.0
smart-open==1.3.0
terminado==0.5
Theano==0.7.0
tornado==4.2.1
tqdm==1.0
traitlets==4.0.0
wheel==0.24.0
